MotionBlockPyramid: compute motion cost in 64-bit to avoid overflow
nLambda and the cost accumulator were widened to int64_t, but
MotionDistorsion truncated the result back to int (`return (int)((nLambda
* dist) >> 8)`), reintroducing the overflow it was meant to avoid: with a
large nLambda (scaled by pixelMax/255 for high bit depth and by nScale^2
at plevel 2) and a large squared vector distance, the value can exceed
INT_MAX and wrap negative, corrupting the `cost >= nMinCost` comparison.
SquareDifferenceNorm likewise computed diff*diff in 32-bit int, which can
overflow for large pel-scaled vectors. Return int64_t from both and do the
squaring/multiply in 64-bit.

+++ src/MotionBlockPyramid.h
@@ -201,7 +201,7 @@ private:
         int pzero, int pglobal, int64_t badSAD, int badrange, bool meander, bool tryMany, bool chroma) noexcept;
 
     bool IsVectorOK(int vx, int vy) const noexcept;
-    int MotionDistorsion(int vx, int vy) const noexcept;
+    int64_t MotionDistorsion(int vx, int vy) const noexcept;
     VECTOR ClipMV(VECTOR v) const noexcept;
     void FetchPredictors(int blkidx, int blkx, int blky, int blkScanDir, VECTOR predictors[5]) noexcept;
     void InitMotionEstimationFields(bool useSatd, bool chroma, int bytesPerSample);
